+++ QuestionsServer/QuestionsData/Queries/QuestionIdQuery.cs
@@ -5,6 +5,6 @@ namespace QuestionsData.Queries
   public class QuestionIdQuery
   {
     [Key]
-    public long QuestionId { get; set; }
+    public int QuestionId { get; set; }
   }
 }
